Control.Kernmantle.Caching

Documentation

data ContentStore

A hash addressed store on the file system.

class Monad m => ContentHashable (m :: Type -> Type) a

type PureHashable = ContentHashable Identity

A value directly hashable

type IOHashable = ContentHashable IO

A value hashable via some IO action

class ProvidesCaching eff where

A class to cache part of the pipeline

Methods

usingStore :: (ContentHashable IO a, Store b) => eff a b -> eff a b

class ProvidesCaching eff => ProvidesPosCaching eff where

A class to cache part of the pipeline where the hash can depend on the position of the task in the pipeline

Methods

usingStore' :: (ContentHashable IO a, Store b) => eff a b -> eff a b

newtype AutoIdent (arr :: Type -> Type -> Type) a b

An arrow transformer that can automatically determine an identifier from its position in a pipeline. It is isomorphic to a Reader ArrowIdent ~> arr, but we need a different Arrow instance than what Cayley provides.

Constructors

AutoIdent (ArrowIdent -> arr a b)

data SomeHashable where

Constructors

SomePureHashable :: PureHashable a => a -> SomeHashable

SomeIOHashable :: IOHashable a => a -> SomeHashable

data StoreWithId remoteCacher

Bundles together a store with an identifier for the whole pipeline. If identifier is Nothing, no caching will be performed.

Constructors

StoreWithId ContentStore remoteCacher (Maybe Int)

type LocalStoreWithId = StoreWithId NoCache

type CachingContext = [SomeHashable]

withStore :: (MonadIO m, MonadMask m) => Path Abs Dir -> (ContentStore -> m a) -> m a

Open the store under the given root and perform the given action. Closes the store once the action is complete

See also: open

caching :: (Arrow core, ProvidesCaching core, ContentHashable IO ident, ContentHashable IO a, Store b) => ident -> Rope r mantle core a b -> Rope r mantle core a b

Any rope whose core provides caching can run cached tasks. The task is identified by an explicit identifier

caching' :: (ProvidesPosCaching core, ContentHashable IO a, Show a, Store b) => Rope r mantle core a b -> Rope r mantle core a b

Any rope whose core provides caching can run cached tasks. The task is identified by its position in the pipeline

localStoreWithId :: ContentStore -> Maybe Int -> LocalStoreWithId

A StoreWithId with no remote caching

Orphan instances

Monad m => ContentHashable m ArrowIdent

Instance details

Methods

contentHashUpdate :: Context SHA256 -> ArrowIdent -> m (Context SHA256)

contentHash :: ArrowIdent -> m ContentHash

Monad m => ContentHashable m SplitId

Instance details

Methods

contentHashUpdate :: Context SHA256 -> SplitId -> m (Context SHA256)

contentHash :: SplitId -> m ContentHash